requirements.txt
netifaces==0.11.0
scapy==2.5.0
paho-mqtt>=1.5.1